# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppStatus = typeof AppStatus[keyof typeof AppStatus];


export const AppStatus = {
  starting: 'starting',
  running: 'running',
  stopped: 'stopped',
  crashed: 'crashed',
} as const;

export interface App {
  id: string;
  name: string;
  code: string;
  port: number;
  status: AppStatus;
  url: string;
  createdAt: string;
}

export interface CreateAppInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
}

export interface AppLogs {
  stdout: string;
  stderr: string;
}

export interface ErrorResponse {
  error: string;
}