# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: apps
    description: Deployed app operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /apps:
    get:
      operationId: listApps
      tags: [apps]
      summary: List all deployed apps
      responses:
        "200":
          description: List of apps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/App"
    post:
      operationId: createApp
      tags: [apps]
      summary: Deploy a new Node.js app
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppInput"
      responses:
        "201":
          description: Deployed app
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /apps/{id}:
    get:
      operationId: getApp
      tags: [apps]
      summary: Get an app by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: App details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteApp
      tags: [apps]
      summary: Stop and delete an app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /apps/{id}/restart:
    post:
      operationId: restartApp
      tags: [apps]
      summary: Restart a deployed app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restarted app
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/App"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /apps/{id}/logs:
    get:
      operationId: getAppLogs
      tags: [apps]
      summary: Get recent logs for an app
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: App logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppLogs"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
    App:
      type: object
      required: [id, name, code, port, status, url, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        port:
          type: integer
        status:
          type: string
          enum: [starting, running, stopped, crashed]
        url:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateAppInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
    AppLogs:
      type: object
      required: [stdout, stderr]
      properties:
        stdout:
          type: string
        stderr:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string