# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppStatus } from './appStatus';

export interface App {
  id: string;
  name: string;
  code: string;
  port: number;
  status: AppStatus;
  url: string;
  createdAt: Date;
}